pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools>=61.2",
    "setuptools_scm",
]

[project]
authors = [{name = "Common Partial Wave Analysis", email = "compwa-admin@ep1.rub.de"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python",
    "Typing :: Typed",
]
dependencies = [
    "PyYAML",
    "attrs >=20.1.0", # https://www.attrs.org/en/stable/changelog.html#id82
    "html2text",
    "ini2toml",
    "nbformat",
    "pip-tools",
    "pre-commit",
    "ruamel.yaml", # better YAML dumping
    "tomlkit",
    'typing-extensions; python_version <"3.10.0"',
]
description = "Pre-commit hooks that ensure that ComPWA repositories have a similar developer set-up"
dynamic = ["version"]
license = {text = "BSD 3-Clause License"}
maintainers = [{email = "compwa-admin@ep1.rub.de"}]
name = "compwa-policy"
requires-python = ">=3.7"

[project.optional-dependencies]
dev = [
    "compwa-policy[doc]",
    "compwa-policy[sty]",
    "compwa-policy[test]",
    "labels",
    "sphinx-autobuild",
    "tox >=1.9", # for skip_install, use_develop
]
doc = [
    "Sphinx",
    "myst-parser",
    "sphinx-api-relink >=0.0.4",
    "sphinx-api-relink",
    "sphinx-argparse",
    "sphinx-book-theme",
    "sphinx-copybutton",
]
format = [
    "black",
]
lint = [
    "compwa-policy[mypy]",
    "radon",
    "ruff",
]
mypy = [
    "mypy",
    "sphinx-api-relink >=0.0.3",
    "types-PyYAML",
    "types-toml",
]
sty = [
    "compwa-policy[format]",
    "compwa-policy[lint]",
    "compwa-policy[test]", # for pytest type hints
    "pre-commit >=1.4.0",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-xdist",
]

[project.readme]
content-type = "text/markdown"
file = "README.md"

[project.scripts]
check-dev-files = "compwa_policy.check_dev_files:main"
colab-toc-visible = "compwa_policy.colab_toc_visible:main"
fix-nbformat-version = "compwa_policy.fix_nbformat_version:main"
format-setup-cfg = "compwa_policy.format_setup_cfg:main"
pin-nb-requirements = "compwa_policy.pin_nb_requirements:main"
self-check = "compwa_policy.self_check:main"
set-nb-cells = "compwa_policy.set_nb_cells:main"

[project.urls]
Source = "https://github.com/ComPWA/policy"
Tracker = "https://github.com/ComPWA/policy/issues"

[tool.setuptools]
include-package-data = false
license-files = ["LICENSE"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
compwa_policy = [
    ".github/*",
    ".github/**/*",
    ".template/*",
    ".template/.*",
    "py.typed",
]

[tool.setuptools.packages.find]
namespaces = false
where = ["src"]

[tool.setuptools_scm]
write_to = "src/compwa_policy/version.py"

[tool.black]
exclude = '''
/(
    .*\.egg-info
    | .*build
    | \.eggs
    | \.git
    | \.pytest_cache
    | \.tox
    | \.venv
    | \.vscode
    | dist
)/
'''
include = '\.pyi?$'
preview = true
target-version = [
    "py310",
    "py311",
    "py312",
    "py37",
    "py38",
    "py39",
]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_defs = true
exclude = '.*/compwa_policy/__init__\.py'
files = ["**/*.py"]
show_error_codes = true
warn_unused_configs = true

[[tool.mypy.overrides]]
check_untyped_defs = true
disallow_incomplete_defs = false
disallow_untyped_defs = false
module = ["tests.*"]

[[tool.mypy.overrides]]
ignore_errors = true
module = ["typings.*"]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = ["ruamel.*"]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = ["nbformat.*"]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = ["pre_commit.commands.autoupdate.*"]

[tool.pyright]
exclude = [
    "**/.git",
    "**/.ipynb_checkpoints",
    "**/.mypy_cache",
    "**/.pytest_cache",
    "**/.tox",
    "**/__pycache__",
    "**/_build",
]
reportGeneralTypeIssues = false
reportIncompatibleMethodOverride = false
reportMissingParameterType = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportOverlappingOverload = false
reportPrivateImportUsage = false
reportPrivateUsage = false
reportUnboundVariable = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryComparison = false
reportUnnecessaryContains = false
reportUnnecessaryIsInstance = false
reportUntypedFunctionDecorator = false
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedImport = true
reportUnusedVariable = true
typeCheckingMode = "strict"

[tool.pytest.ini_options]
addopts = [
    "--color=yes",
    "--doctest-continue-on-failure",
    "--doctest-modules",
    "--durations=3",
]
filterwarnings = [
    "error",
    "ignore: Importing ErrorTree directly from the jsonschema package is deprecated.*",
]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
testpaths = [
    "src",
    "tests",
]

[tool.ruff]
extend-exclude = ["typings"]
extend-select = [
    "A",
    "B",
    "BLE",
    "C4",
    "C90",
    "D",
    "EM",
    "ERA",
    "FA",
    "I",
    "ICN",
    "INP",
    "ISC",
    "N",
    "NPY",
    "PGH",
    "PIE",
    "PL",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "T20",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
ignore = [
    "C408",
    "D101",
    "D102",
    "D103",
    "D105",
    "D107",
    "D203",
    "D213",
    "D407",
    "D416",
    "E501",
    "PLR0913",
    "PLW2901",
    "S301",
    "SIM108",
    "UP036",
]
show-fixes = true
src = [
    "src",
    "tests",
]
target-version = "py37"
task-tags = ["cspell"]

[tool.ruff.isort]
known-first-party = ["compwa_policy"]

[tool.ruff.per-file-ignores]
"docs/*" = [
    "E402",
    "INP001",
    "S101",
    "S113",
    "T201",
]
"docs/conf.py" = ["D100"]
"setup.py" = ["D100"]
"tests/*" = [
    "D",
    "INP001",
    "PGH001",
    "PLR0913",
    "PLR2004",
    "S101",
    "T20",
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.tomlsort]
all = false
ignore_case = true
in_place = true
sort_first = [
    "build-system",
    "project",
    "tool.setuptools",
    "tool.setuptools_scm",
]
sort_table_keys = true
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
